% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfan_says.R
\name{irfan_says}
\alias{irfan_says}
\title{Irfan says}
\usage{
irfan_says(saying = NULL)
}
\arguments{
\item{saying}{\code{character}. You can specify, which saying you'd like
to be printed. By default (\code{NULL}), it takes a random sample.}
}
\description{
Irfan says
}
\examples{
 irfan_says(saying = "coffee")

}
\author{
Christoph Stepper
}
